NPM_total

A function to calculate total NPM score Logic for this function is documented in Nutrient Profiling Technical Guidance page 6.

Usage

NPM_total(a_score, c_score, fvn_score, fibre_score)

Arguments

a_score,

a numeric value for the A score

c_score,

a numeric value for the C score

fvn_score,

a numeric value for the specific score for fruit/veg/nuts percentage

fibre_score,

a numeric value for the fibre score

Value

a numeric value for the overall NPM score