Example dataset called CDRC Drinks

A handcrafted set of data that can be used as the input for nutrientprofiler functions

Usage

cdrcdrinks

Format

cdrcdrinks ## A data frame of 10 rows and 21 columns:

name

Name of product

brand

Brand of product

product_category

Category of product

product_type

Type of product, i.e. Food or Drink

food_type

Type of food, used for specific gravity conversion

drink_format

Format of drink, used for specific gravity conversion

drink_type

Type of drink, used for speciifc gravity conversion

nutrition_info

How nutritional information is presented, for specific gravity conversion

energy_measurement_kj

Energy measurement in KJ

energy_measurement_kcal

Energy measurement in kcal

sugar_measurement_g

Sugar content measured in in grams

fat_measurement_g

The fat content measured in grams

salt_measurement_g

The salt content measured in grams

sodium_measurement_mg

The salt content measured in mg of sodium

fibre_measurement_nsp

The fibtre measurement using NSP

fibre_measurement_aoac

The fibre measurement using AOAC

protein_measurement_g

The protein emasurement of the product

fruit_nut_measurement_percent

The percent of fruit and nuts in the product

weight_g

The mass of the product if solid

volume_ml

The volume of the product if liquid

volume_water_ml

The volume of water for diluted products